+++ changelog.md
@@ -1,7 +1,9 @@
 # Release 3.0.0-SNAPSHOT
 1. Refactored code to fit better into libGdx/packr parent repository.
 1. Fixed an issue where extracting an archive with duplicate entries would fail.
-1. The packr-all Jar is published to GitHub packages <https://github.com/libgdx/packr/packages>.
+1. The packr-all Jar is available from GitHub packages <https://github.com/libgdx/packr/packages>.
+1. The output directory specified by `--output` must be an empty directory, or a path that does not exist.
+   * Packr will no longer delete the output directory and then populate it.
 # Release 2.7.0
 1. Fixed a Gradle script error where it was bundling the release builds with debug info on Linux and macOS.
    * For Linux this reduces the executable size from ~722K to ~95K.